changed structure of views into files
