+++ calendar_app/templates/calendar_app/calendar_month.html
@@ -36,7 +36,7 @@ td {
                 <a class="btn btn-primary disabled" href="{% url 'month-view' %}" role="button">Month View</a>
                 <a class="btn btn-primary" href="{% url 'week-view' %}" role="button">Week View</a>
                 <!-- Task creation button -->
-                <a class="btn btn-primary" href="{% url 'create_task' %}" role="button">Add Task</a>
+                <a class="btn btn-primary" href="{% url 'create-task' %}" role="button">Add Task</a>
                 <div class ="dropdown btn btn-primary">
                     <!-- Dropdown for the different categories-->
                     <button class="btn btn-primary dropdown-toggle " href="#" role="button" data-bs-toggle="dropdown">Filter</button>
